backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
pytest
httpx
